optimize: op_info.xrst: change addr_t -> index_t. list_setvec.hpp: set is_pod for all possible optimizer types.

+++ include/cppad/local/play/random_iterator.hpp
@@ -2,7 +2,7 @@
 # define CPPAD_LOCAL_PLAY_RANDOM_ITERATOR_HPP
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-or-later
 // SPDX-FileCopyrightText: Bradley M. Bell <bradbell@seanet.com>
-// SPDX-FileContributor: 2003-24 Bradley M. Bell
+// SPDX-FileContributor: 2003-25 Bradley M. Bell
 // ----------------------------------------------------------------------------
 
 // BEGIN_CPPAD_LOCAL_PLAY_NAMESPACE
@@ -102,7 +102,7 @@ public:
       size_t&        var_index  ) const
    {  op        = op_code_var( (*op_vec_)[op_index] );
       op_arg    = (*op2arg_vec_)[op_index] + arg_vec_->data();
-      var_index = (*op2var_vec_)[op_index];
+      var_index = size_t( (*op2var_vec_)[op_index] );
       return;
    }
    /*!
